app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.12
aiofiles==24.1.0
groq==0.13.0
python-dotenv==1.0.1
pandas==2.2.3
bcrypt==4.2.0
itsdangerous==2.2.0
psycopg[binary]>=3.3.0
